requirements.txt
Flask==3.0.0
opencv-python-headless==4.8.1.78
numpy==1.26.2
git+https://github.com/ageitgey/face_recognition_models
face_recognition==1.3.0
setuptools<70
gunicorn==21.2.0
python-dotenv==1.0.0
pg8000==1.30.5
